queries/txns: include TXN-less accesses as well
This query deals with memory accesses outside of TXNs differently: Instead
of folding them all into one single TXN, it acts as if each was carried out
in a TXN of its own.  This makes sense, because otherwise even a large
number of accesses to a particular member would be folded into a single
access and a single TXN, making it impossible to determine its frequency
(and, hence, the probability that this member may really be accessed
without any locks held) afterwards.