hypothesizer: prefer hypotheses with more locks
If there are hypotheses with the same occurrence count, prefer those with
more locks.  For example:

100% (3004 out of 3004 mem accesses under locks):
  EMBSAME(i_mutex) -> EMBSAME(i_data.i_mmap_mutex)
100% (3004 out of 3004 mem accesses under locks):
  EMBSAME(i_data.i_mmap_mutex)
100% (3004 out of 3004 mem accesses under locks):
  EMBSAME(i_mutex)

In this example, the A->B locking rule naturally implies that A and B alone
also pose valid locking rules with at least the observation coverage of
A->B.  But if the observation coverage is the same (100% in the example),
it is safe to assume that A->B is the correct rule.