hypothesizer: properly sort by member name
Before this change, hypothesizer sorted by member name including the
access-type prefix (r: / w:), hence r:dev would appear far away from w:dev
in the output.  As a kernel developer pointed out this is inconvenient,
therefore the access type and member name is now stored (and sorted)
separately.

Note that the CSV output still shows the "combined" name including the
access type.  This could now easily be changed in to two separate CSV
columns.

An open TODO is to properly use this separation in the source-code comment
output, e.g., use it for sorting there as well, or even merge read and
write hypotheses if the corresponding locking rule is identical.