hypothesizer: less confusing hypothesis list
This commit is intended to reduce confusion in the hypothesis list,
which in the CSV variant even included duplicates.  In the case that
lock combination observations contained a specific lock multiple
times, you could see output like this:

 3.25% (264 out of 8115 mem accesses under locks): EMBOTHER(bd_mutex)
    95.8% EMBOTHER(bd_mutex)
    4.17% EMBOTHER(bd_mutex) -> EMBOTHER(bd_mutex)

Later in the output, you would see "EMBOTHER(bd_mutex) ->
EMBOTHER(bd_mutex)" as a separate hypothesis again (and both
occurrences would look identical in the CSV output).  The output
actually makes sense: EMBOTHER(bd_mutex) was seen in 264 mem accesses;
in 95.8% of these it was held once, in 4.17% it was held twice.

But as this information is rarely useful and mostly confusing, this
commit makes sure multiple occurrences of one specific lock only show
up once in the output, simplifying the above example to:

 3.25% (264 out of 8115 mem accesses under locks): EMBOTHER(bd_mutex)