conv-import.sh: Added optional variable PROCESS_PREEMPT_COUNT
If PROCESS_PREEMPT_COUNT is set (>0) in convert.conf, convert will be
started with -p. This will enable processing of the hard- and
softirq lock based on the preempt_count.
If the kernel was compiled with CONFIG_ESS_LOCK_ANALYSIS_TRACE_*,
PROCESS_PREEMPT_COUNT should be zero to use the hard- and softirq
lock.