Generate a pretty bug overview
get-counterexample.sh just runs the queries retrieving the cex. It
does not resolve any instruction pointer anymore. pretty-print-cex.py uses the output
of the former script to generate a pretty html site which illustrates the counterexamples.
That includes links to the source code of stack entries as well as the
source code where locks were acquired.