Resolve a global lock's variable name
Convert resolves a global lock's address to a variable name, which
is stored in the database. For those global locks that are embedded in
a data structure the respective variable name is used.
The following post processing uses that name as well. Instead of
just using a lock's id to form an unique name, e.g., 4711(spinlock_t),
its variable name is now used: console_sem:3(semaphore).