Introduced a new lock model (1/2)
We now fully support reader, writer, and reader-writer locks.
Each lock is represented by two so-called sub locks: the
reader, and the writer sub lock. For reader-only and writer-only
locks,
the writer and the reader part is not used, respectively.
To distinguish both sub locks in the post processing stage,
each sub lock is individually stored in the databse.
In post processing, "[r]" or "[w]" is appended to a lock's name
for the respective sub locks.

This commit contains the changes made to convert.