Use a smoother step fn to calculate the confidence
For sample sizes per hypothesis above the confidence threshold (param. -c)
the confidence is equal to the relative support.
Below the confidence threshold, the relative support is multiplied by
the value of the smother step function. The input for that function
is the absolute support divided by the confidence threshold.