Strip down the stacktrace and store its elements in the DB
Currently the function blacklist just effects the functions
where a memory access happens.
However, the initialization might happen somewhere deeper
in the code, i.e., a blacklisted function calls
a specialized function which does the actual access.
This case is currently ignored.
We therefore strip down the stacktrace and store each stackframe
in the database.
We can now apply the function blacklist to each stackframe, and
filter memory accesses that lie on an init path.