Resolve inlined functions for tigger IP
Only resolve potential inlined functions for the* instruction pointer
that triggered the memory access (--> sequence == 0).
The instruction pointers found on the stack are return addresses.
That means each ip indicates the *next* instruction to be executed.
Resolving that ip for inlined functions may lead to a corrupt stacktrace
which includes functions that have not been called yet.