hypothesizer: new winner-selection strategy
Our producer/consumer toy example showed that the previous
winner-selection strategy -- to pick the hypothesis with the most
relative support -- fails even in simple cases.  In this scenario, a
specific member had to be accessed with two locks A and B held; due to
a synthetic "bug" in the locking code, in one instance the member was
accessed with only A held.  For an assumed total number of accesses N
to this member, the following hypotheses were generated:

 A -> B, support = N-1
      A, support = N
      B, support = N-1

With the previous strategy, the "A" locking rule was mistakenly
chosen.

The new strategy implemented in this commit instead picks the
locking-rule hypothesis with the *lowest* support that exceeds the
accept threshold; if there are multiple hypotheses with the same
support, the one with the most locks is chosen.  This prevents
"overfitting" the generated locking rule to the observed data, as it
can deal with a few wrongly locked memory accesses.

As a side-effect, the "no locks needed" case is no longer a special
case (besides mainly output-formatting details); it is handled as a
locking-rule hypothesis like all others, and always ends up at the top
of the list with 100% support and the lowest number (0) of locks.  It
gets chosen if no other locking rule exceeds the accept threshold; the
-n/--nolock-threshold parameter does not exist anymore.