Introduced subclasses of data types
Since kernel developers tend to optimize their code
for performance, they want to reduce locking to a bare minimum.
They therefore use as much knowledge about of a data type as possible.
Locking an inode, for example, is tailored to the filesystem that inode stands for.
This can lead to less locking compared to a generic inode
if, for example, some operations are not supported by that filesystem.
Hence, a filesystem's inode represents a subclass of the generic inode.
Deriving locking rules from observations of different subclasses
can lead to wrong locking rules or false positive counterexamples.
We've now introduced subclasses to LockDoc.
We can now distinguish between ext4 or procfs inodes, for example,
and derive locking rules for them separately.
This concept does not only apply for inodes. It is applicable for all sort
of data types that have subclasses.