convert: Use a heuristic to resolve code locations
On x86, the return address points to the instruction to be xecuted
*next*. Resolving that instruction pointer including potential
inlined functions may lead to fake stacktrace entries.
Those entries point to future function calls. We therefore decrement the return address by one before resolving
it. This creates an address pointing into the previous instruction
aka a call instruction. Libbfd is able to resolve addresses that are
not aligned to instruction boundaries.