convert: Improved name lookup of global locks
Dwarves fails to resolve a global lock's variable name if it is
non-static, i.e. visible in other cus.
This is due to the fact that the DWARF info do not contain any
variable definition for such a lock. They just contain variable declarations.
convert now queries the global symbol table for such locks
if it failed in the first place.