function blacklist: Added filter option 'sequence'
The function blacklist is applied to the whole stacktrace.
Therefore, each access performed within a blacklisted function
is ignored. This is true even if the blacklisted function is
the root of the call trace, and the access is *not* performed
within the function itself.
Some times we just want to ignore accesses that are executed
directly in a function, but not those performed in called
functions.
This can now be achieved by specifying a sequence number for
each entry of the function blacklist.
NULL (or \N) means an entry is applied to the whole stack trace.
Whereas 0-n correspond to the n-th element of the trace.
Zero means the topmost element.