convert: Fake array size for zero-sized arrays, e.g. char name[0].
Developers often want to have structs with
elements whose size is determined at allocation time.
They therefore use patterns like this:
struct foo {
  ...
  char bar[0]
};
Since a ptr of struct foo usually points to the beginning
of the memory area, the remaining memory can be accessed
via bar.
This patterns makes it impossible to map memory accesses to
that area to an element of struct foo.
We therefore fake the size of array bar. For us, it will be 1MiByte.
1MiByte should be enough for everyone. :)