convert: Fixed sign problems with getopt()
On at least aarch64, char is internally an unsigned type.
Thus, -1 returned by getopt() is converted to 255.
It is then compared to -1 in the condition of our loop
which is always false. This leads to an endless loop.
We now store the return value of getopt() in an int
variable.
This preserves the -1 as well as allows the comparison
with characters.